online quiz system
Today Online Quiz System has become a fast-growing examination method because of its speed and accuracy. It also needed less manpower to execute the examination. Almost all organizations nowadays are conducting their objective exams by online examination or quiz system, which saves students time in examinations. Organizations can also easily check the performance of the students that they take in an examination. As a result of this, organizations are releasing results in less time. It also helps the environment by saving paper. According to today’s requirements, an online examination project in PHP is very useful to learn it. So, I have made a full stack project on the online quiz system.